+++ examples/with_vertexai.py
@@ -1,13 +1,16 @@
 """Example of using PandasAI with a CSV file and Google Vertexai."""
 
+import os
 import pandas as pd
-
 from pandasai import SmartDataframe
-from pandasai.llm import GoogleVertexai
+from pandasai.llm import GoogleVertexAI
 
 df = pd.read_csv("examples/data/Loan payments data.csv")
 
-llm = GoogleVertexai(
+# Set the path of your json credentials
+os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = "application_default_credentials.json"
+
+llm = GoogleVertexAI(
     project_id="generative-ai-training", location="us-central1", model="text-bison@001"
 )
 df = SmartDataframe(df, config={"llm": llm})
