+++ tests/llms/test_google_vertexai.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from pandasai.llm import GoogleVertexai
+from pandasai.llm import GoogleVertexAI
 
 
 class MockedCompletion:
@@ -10,11 +10,11 @@ class MockedCompletion:
         self.result = result
 
 
-class TestGoogleVertexai:
+class TestGoogleVertexAI:
     def test_init_with_default_model(self):
         project_id = "your_project_id"
         location = "northamerica-northeast1"
-        vertexai_instance = GoogleVertexai(project_id, location)
+        vertexai_instance = GoogleVertexAI(project_id, location)
 
         assert vertexai_instance.model == "text-bison@001"
         assert vertexai_instance.project_id == project_id
@@ -25,7 +25,7 @@ class TestGoogleVertexai:
         location = "northamerica-northeast1"
         custom_model = "code-bison@001"
 
-        vertexai_instance = GoogleVertexai(project_id, location, model=custom_model)
+        vertexai_instance = GoogleVertexAI(project_id, location, model=custom_model)
 
         assert vertexai_instance.model == custom_model
         assert vertexai_instance.project_id == project_id
@@ -37,13 +37,13 @@ class TestGoogleVertexai:
         project_id = "test-project"
         location = "northamerica-northeast1"
         custom_model = "code-bison@001"
-        return GoogleVertexai(project_id, location, custom_model)
+        return GoogleVertexAI(project_id, location, custom_model)
 
-    def test_validate_with_model(self, google_vertexai: GoogleVertexai):
+    def test_validate_with_model(self, google_vertexai: GoogleVertexAI):
         google_vertexai.model = "text-bison@001"
         google_vertexai._validate()  # Should not raise any errors
 
-    def test_validate_without_model(self, google_vertexai: GoogleVertexai):
+    def test_validate_without_model(self, google_vertexai: GoogleVertexAI):
         google_vertexai.model = None
         with pytest.raises(ValueError, match="model is required."):
             google_vertexai._validate()
